fix: Correct sql
- 잘못된 sql문 수정

issue: #7
