+++ be/src/main/java/com/team18/baseball/entity/BattingResult.java
@@ -1,4 +1,4 @@
-package com.team18.baseball.domain;
+package com.team18.baseball.entity;
 
 public class BattingResult {
     private final BattingResultType battingResultType;
