+++ be/src/main/java/com/team18/baseball/entity/PlateAppearanceInfo.java
@@ -1,4 +1,4 @@
-package com.team18.baseball.domain;
+package com.team18.baseball.entity;
 
 public class PlateAppearanceInfo {
     private final Long id;
