+++ be/src/main/java/com/team18/baseball/entity/Team.java
@@ -1,4 +1,4 @@
-package com.team18.baseball.domain;
+package com.team18.baseball.entity;
 
 import org.springframework.data.annotation.Id;
 
@@ -6,9 +6,9 @@ import java.util.List;
 
 public class Team {
     @Id
-    private final Long id;
+    private Long id;
     private final String name;
-    private final List<Player> players;
+    private List<Player> players;
 
     Team(Long id, String name, List<Player> players) {
         this.id = id;
@@ -18,5 +18,9 @@ public class Team {
 
     public static Team create(Long id, String name, List<Player> players) {
         return new Team(id, name, players);
+    }
+
+    public Long getId() {
+        return id;
     }
 }
