+++ be/src/main/java/com/team18/baseball/repository/TeamRepository.java
@@ -1,0 +1,7 @@
+package com.team18.baseball.repository;
+
+import com.team18.baseball.entity.Team;
+import org.springframework.data.repository.CrudRepository;
+
+public interface TeamRepository extends CrudRepository<Team, Long> {
+}
